# perf(utils): add benchmarks for file system walk and glob libs