--- a/package.json
+++ b/package.json
@@ -67,6 +67,7 @@
     "@commitlint/config-nx-scopes": "^18.4.3",
     "@commitlint/cz-commitlint": "^17.7.1",
     "@jscutlery/semver": "^4.1.0",
+    "@nodelib/fs.walk": "^2.0.0",
     "@nx/devkit": "17.1.3",
     "@nx/esbuild": "17.1.3",
     "@nx/eslint-plugin": "17.1.3",
@@ -110,6 +111,9 @@
     "eslint-plugin-sonarjs": "^0.22.0",
     "eslint-plugin-unicorn": "^48.0.1",
     "eslint-plugin-vitest": "^0.3.8",
+    "fast-glob": "^3.3.2",
+    "glob": "^10.3.10",
+    "globby": "^14.0.1",
     "husky": "^8.0.0",
     "inquirer": "^8.2.6",
     "jsdom": "~22.1.0",
@@ -128,11 +132,7 @@
     "verdaccio": "^5.0.4",
     "vite": "~4.3.9",
     "vitest": "~0.32.0",
-    "zod2md": "^0.1.1",
-    "@nodelib/fs.walk": "^2.0.0",
-    "fast-glob": "^3.3.2",
-    "glob": "^10.3.10",
-    "globby": "^14.0.1"
+    "zod2md": "^0.1.1"
   },
   "optionalDependencies": {
     "@esbuild/darwin-arm64": "^0.19.4",