--- a/packages/utils/perf/crawl-file-system/optimized0.ts
+++ b/packages/utils/perf/crawl-file-system/optimized0.ts
@@ -1,8 +1,0 @@
-import { CrawlFileSystemOptions } from '../../src';
-
-export function crawlFileSystemOptimized0<T = string>(
-  options: CrawlFileSystemOptions<T>,
-): Promise<T[]> {
-  const { directory, pattern: needle = '@TODO' } = options;
-  return Promise.resolve([directory, needle] as T[]);
-}