--- a/packages/utils/perf/glob/fast-glob.ts
+++ b/packages/utils/perf/glob/fast-glob.ts
@@ -1,0 +1,5 @@
+import * as fg from 'fast-glob';
+
+export function fastGlob(pattern: string[]): Promise<string[]> {
+  return fg.async(pattern);
+}