--- a/packages/utils/perf/glob/glob.ts
+++ b/packages/utils/perf/glob/glob.ts
@@ -1,0 +1,5 @@
+import { glob as g } from 'glob';
+
+export function glob(pattern: string[]): Promise<string[]> {
+  return g(pattern);
+}