--- a/packages/utils/perf/glob/globby.ts
+++ b/packages/utils/perf/glob/globby.ts
@@ -1,0 +1,5 @@
+import { globby as g } from 'globby';
+
+export function globby(pattern: string[]): Promise<string[]> {
+  return g(pattern);
+}