This PR includes perf benchmarks for:
- file system crawling
  - [@nodelib/fs.walk](https://www.npmjs.com/package/@nodelib/fs.walk)
- glob search libs
  - [fast-glob](https://www.npmjs.com/package/fast-glob)
  - [glob](https://www.npmjs.com/package/glob)
  - [globby](https://www.npmjs.com/package/globby)

---

**file system crawling**  

```bash
> nx run utils:perf crawl-file-system

You can adjust the test with the following arguments:
directory      target directory of test      --directory=/Users/michael_hladky/WebstormProjects/quality-metrics-cli/node_modules
pattern        pattern to search             --pattern=/.json$/
 
Start benchmark...
 
Base x 2.43 ops/sec ±1.45% (17 runs sampled)
nodelib.fsWalk x 17.09 ops/sec ±0.74% (82 runs sampled)
 
Total Duration: 13.08 sec
Fastest is nodelib.fsWalk
```

**glob search libs**  
```bash
> nx run utils:perf glob

You can adjust the test with the following arguments:
pattern      glob pattern of test      --pattern=/Users/michael_hladky/WebstormProjects/quality-metrics-cli/node_modules/**/*.js
 
Start benchmark...
 
glob found 27167 files
glob x 11.66 ops/sec ±19.61% (67 runs sampled)
globby found 27164 files
globby x 9.18 ops/sec ±0.74% (48 runs sampled)
fastGlob found 27164 files
fastGlob x 12.62 ops/sec ±0.76% (63 runs sampled)
 
Total Duration: 18.21 sec
Fastest is fastGlob
```